refactor: create token.rs
